loosen bundler restriction
I don't know if this is a good idea. First I want to see if this works.
I don't know enough about Travis yet to know why, but `bundle install`
was run before the before_script, so the tests died on 2.5/2.6 which
default to bundler 1.17.

This is a gem, so we don't commit the Gemfile.lock, and I'm not aware of
any specific bundler configuration that might make this gem incompatible
with the older bundler. My goal would be for this gem to just work for
people regardless of bundler version, so if I can avoid the restriction,
I'd like to.